Ignore RLIMIT_RSS on Apple platforms.
On Apple platforms, RLIMIT_RSS is mapped to RLIMIT_AS [1] and setting
RLIMIT_AS to a value smaller than the current virtual memory size
will fail, This is incompatible with the current usage in timeit
and can cause issues on platforms enforcing strict virtual memory
size limits. Ignore RLIMIT_RSS on Apple platforms for now.

See D119572 for a possible alternative, but this comes with its own
issues, in that the default value passed by llvm-lit is not sufficiently
large and some benchmarks need larger virtual memory.

[1] https://github.com/apple/darwin-xnu/blob/a1babec6b135d1f35b2590a1990af3c5c5393479/bsd/sys/resource.h